# docs/swagger.yaml
basePath: /api
definitions:
  subscription.ErrorResponse:
    properties:
      error:
        example: invalid date format, use MM-YYYY
        type: string
    type: object
  subscription.ListResult:
    properties:
      has_more:
        example: true
        type: boolean
      limit:
        example: 50
        type: integer
      offset:
        example: 10
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/subscription.SubscriptionResp'
        type: array
      total:
        example: 1000
        type: integer
    type: object
  subscription.SubscriptionReq:
    properties:
      end_date:
        example: 12-2026
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  subscription.SubscriptionResp:
    properties:
      created_at:
        example: "2025-02-19T10:00:00Z"
        type: string
      end_date:
        example: 12-2026
        type: string
      id:
        example: 42
        type: integer
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      updated_at:
        example: "2025-02-19T10:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  subscription.TotalCostResp:
    properties:
      end_date:
        example: 12-2025
        type: string
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 01-2025
        type: string
      subscriptions_count:
        example: 4
        type: integer
      total_cost:
        example: 4800
        type: integer
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - end_date
    - start_date
    type: object
host: localhost:8080
info:
  contact:
    email: tagwaylander223@gmail.com
    name: Oleg Bogdanov
  description: |-
    REST API service for managing user subscriptions
    Test task for Effective Mobile
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get paginated list of all subscriptions
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.ListResult'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: List all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription for a user
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subscription.SubscriptionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscription.SubscriptionResp'
        "400":
          description: Invalid request body or date format
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "404":
          description: Subscription not found
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.SubscriptionResp'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "404":
          description: Subscription not found
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update an existing subscription (full update)
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subscription.SubscriptionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.SubscriptionResp'
        "400":
          description: Invalid request or date format
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "404":
          description: Subscription not found
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total-cost:
    get:
      consumes:
      - application/json
      description: Calculate total cost of all subscriptions in a given period with
        optional filters
      parameters:
      - description: Start date (MM-YYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      - description: Filter by user ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.TotalCostResp'
        "400":
          description: Invalid parameters or date format
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Calculate total cost of subscriptions
      tags:
      - subscriptions
swagger: "2.0"