# internal/subscription/handlers.go
package subscription

import (
	"context"
	"errors"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

const requestTimeout = 5 * time.Second

type Handler struct {
	service *Service
	logger  *zap.Logger
}

func RegisterRoutes(rg *gin.RouterGroup, service *Service, logger *zap.Logger) {
	h := &Handler{service: service, logger: logger}
	rg.POST("/subscriptions", h.CreateSubscription)
	rg.GET("/subscriptions/:id", h.GetSubscription)
	rg.PUT("/subscriptions/:id", h.UpdateSubscription)
	rg.DELETE("/subscriptions/:id", h.DeleteSubscription)
	rg.GET("/subscriptions", h.ListSubscriptions)

	rg.GET("/subscriptions/total-cost", h.GetTotalCost)
}

// CreateSubscription godoc
// @Summary      Create a new subscription
// @Description  Create a new subscription for a user
// @Tags         subscriptions
// @Accept       json
// @Produce      json
// @Param        subscription body SubscriptionReq true "Subscription data"
// @Success      201  {object}  SubscriptionResp
// @Failure      400  {object}  ErrorResponse  "Invalid request body or date format"
// @Failure      500  {object}  ErrorResponse  "Internal server error"
// @Router       /subscriptions [post]
func (h *Handler) CreateSubscription(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), requestTimeout)
	defer cancel()

	var req SubscriptionReq
	if err := c.ShouldBindJSON(&req); err != nil {
		h.logger.Warn("invalid request body",
			zap.Int("status", 400),
			zap.String("error", err.Error()),
			zap.String("path", c.Request.URL.Path),
			zap.String("method", c.Request.Method))
		c.AbortWithStatusJSON(400, gin.H{"error": "invalid request body"})
		return
	}

	h.logger.Debug("creating subscription",
		zap.String("user_id", req.UserID.String()),
		zap.String("service", req.ServiceName),
		zap.Int("price", req.Price))

	response, err := h.service.CreateSubscription(ctx, req)
	if err != nil {
		switch {
		case errors.Is(err, context.DeadlineExceeded):
			c.JSON(504, gin.H{"error": "timeout"})
		case errors.Is(err, ErrInvalidDateFormat), errors.Is(err, ErrEndDateBeforeStart):
			c.JSON(400, gin.H{"error": err.Error()})
		default:
			h.logger.Error("failed to create subscription",
				zap.Int("status", 500),
				zap.String("error", err.Error()),
				zap.String("path", c.Request.URL.Path))
			c.JSON(500, gin.H{"error": "internal error"})
		}
		return
	}
	h.logger.Info("subscription created successfully",
		zap.String("user_id", req.UserID.String()),
		zap.String("service", req.ServiceName),
		zap.Int("price", req.Price))

	c.JSON(201, response)
}

// GetSubscription godoc
// @Summary      Get subscription by ID
// @Description  Get detailed information about a specific subscription
// @Tags         subscriptions
// @Accept       json
// @Produce      json
// @Param        id   path      int  true  "Subscription ID"
// @Success      200  {object}  SubscriptionResp
// @Failure      400  {object}  ErrorResponse  "Invalid ID format"
// @Failure      404  {object}  ErrorResponse  "Subscription not found"
// @Failure      500  {object}  ErrorResponse  "Internal server error"
// @Router       /subscriptions/{id} [get]
func (h *Handler) GetSubscription(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), requestTimeout)
	defer cancel()

	id, err := getID(c)
	if err != nil {
		h.logger.Warn("invalid id format",
			zap.String("id", c.Param("id")),
			zap.Error(err))
		c.JSON(400, gin.H{"error": "invalid subscription id"})
		return
	}

	h.logger.Debug("getting subscription",
		zap.Int64("subscription_id", id))

	response, err := h.service.GetSubscription(ctx, id)
	if err != nil {
		switch {
		case errors.Is(err, context.DeadlineExceeded):
			c.JSON(504, gin.H{"error": "timeout"})
		case errors.Is(err, ErrSubscriptionNotFound):
			c.JSON(404, gin.H{"error": "subscription not found"})
		default:
			h.logger.Error("failed to get subscription",
				zap.Int("status", 500),
				zap.String("error", err.Error()),
				zap.String("path", c.Request.URL.Path))
			c.JSON(500, gin.H{"error": "internal error"})
		}
		return
	}
	h.logger.Info("subscription retrieved successfully",
		zap.Int64("id", id))

	c.JSON(200, response)
}

// UpdateSubscription godoc
// @Summary      Update subscription
// @Description  Update an existing subscription (full update)
// @Tags         subscriptions
// @Accept       json
// @Produce      json
// @Param        id            path      int             true  "Subscription ID"
// @Param        subscription body     SubscriptionReq true  "Updated subscription data"
// @Success      200  {object}  SubscriptionResp
// @Failure      400  {object}  ErrorResponse  "Invalid request or date format"
// @Failure      404  {object}  ErrorResponse  "Subscription not found"
// @Failure      500  {object}  ErrorResponse  "Internal server error"
// @Router       /subscriptions/{id} [put]
func (h *Handler) UpdateSubscription(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), requestTimeout)
	defer cancel()

	id, err := getID(c)
	if err != nil {
		h.logger.Warn("invalid id format",
			zap.String("id", c.Param("id")),
			zap.Error(err))
		c.JSON(400, gin.H{"error": "invalid subscription id"})
		return
	}

	var req SubscriptionReq
	if err := c.ShouldBindJSON(&req); err != nil {
		h.logger.Warn("invalid request body",
			zap.Int("status", 400),
			zap.String("error", err.Error()),
			zap.String("path", c.Request.URL.Path),
			zap.String("method", c.Request.Method))
		c.AbortWithStatusJSON(400, gin.H{"error": "invalid request body"})
		return
	}

	h.logger.Debug("updating subscription",
		zap.Int64("subscription_id", id))

	response, err := h.service.UpdateSubscription(ctx, id, req)
	if err != nil {
		switch {
		case errors.Is(err, context.DeadlineExceeded):
			c.JSON(504, gin.H{"error": "timeout"})
		case errors.Is(err, ErrInvalidDateFormat), errors.Is(err, ErrEndDateBeforeStart):
			c.JSON(400, gin.H{"error": err.Error()})
		case errors.Is(err, ErrSubscriptionNotFound):
			c.JSON(404, gin.H{"error": "subscription not found"})
		default:
			h.logger.Error("failed to update subscription",
				zap.Int("status", 500),
				zap.String("error", err.Error()),
				zap.String("path", c.Request.URL.Path))
			c.JSON(500, gin.H{"error": "internal error"})
		}
		return
	}

	h.logger.Info("subscription updated successfully",
		zap.Int64("id", id))

	c.JSON(200, response)
}

// DeleteSubscription godoc
// @Summary      Delete subscription
// @Description  Delete a subscription by ID
// @Tags         subscriptions
// @Accept       json
// @Produce      json
// @Param        id   path      int  true  "Subscription ID"
// @Success      204  "No Content"
// @Failure      400  {object}  ErrorResponse  "Invalid ID format"
// @Failure      404  {object}  ErrorResponse  "Subscription not found"
// @Failure      500  {object}  ErrorResponse  "Internal server error"
// @Router       /subscriptions/{id} [delete]
func (h *Handler) DeleteSubscription(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), requestTimeout)
	defer cancel()

	id, err := getID(c)
	if err != nil {
		h.logger.Warn("invalid id format",
			zap.String("id", c.Param("id")),
			zap.Error(err))
		c.JSON(400, gin.H{"error": "invalid subscription id"})
		return
	}
	h.logger.Debug("deleting subscription",
		zap.Int64("subscription_id", id))

	err = h.service.DeleteSubscription(ctx, id)
	if err != nil {
		switch {
		case errors.Is(err, context.DeadlineExceeded):
			c.JSON(504, gin.H{"error": "timeout"})
		case errors.Is(err, ErrSubscriptionNotFound):
			c.JSON(404, gin.H{"error": "subscription not found"})
		default:
			h.logger.Error("failed to delete subscription",
				zap.Int("status", 500),
				zap.String("error", err.Error()),
				zap.String("path", c.Request.URL.Path))
			c.JSON(500, gin.H{"error": "internal error"})
		}
		return
	}
	h.logger.Info("subscription deleted successfully",
		zap.Int64("id", id))

	c.Status(204)
}

// ListSubscriptions godoc
// @Summary      List all subscriptions
// @Description  Get paginated list of all subscriptions
// @Tags         subscriptions
// @Accept       json
// @Produce      json
// @Param        offset query int false "Offset for pagination" default(0) minimum(0)
// @Param        limit  query int false "Limit for pagination" default(10) minimum(1) maximum(100)
// @Success      200  {object}  ListResult
// @Failure      500  {object}  ErrorResponse  "Internal server error"
// @Router       /subscriptions [get]
func (h *Handler) ListSubscriptions(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), requestTimeout)
	defer cancel()

	offset, err := strconv.Atoi(c.DefaultQuery("offset", "0"))
	if err != nil {
		h.logger.Warn("invalid offset",
			zap.String("offset", c.Query("offset")),
			zap.Error(err))
		c.JSON(400, gin.H{"error": "invalid offset"})
		return
	}

	if offset < 0 {
		offset = 0
	}

	limit, err := strconv.Atoi(c.DefaultQuery("limit", "10"))
	if err != nil {
		h.logger.Warn("invalid limit",
			zap.String("limit", c.Query("limit")),
			zap.Error(err))
		c.JSON(400, gin.H{"error": "invalid limit"})
		return
	}

	if limit <= 0 {
		limit = 10
	}

	if limit > 100 {
		limit = 100
	}

	h.logger.Debug("listing subscriptions",
		zap.Int("offset", offset),
		zap.Int("limit", limit))

	response, err := h.service.ListSubscriptions(ctx, offset, limit)
	if err != nil {
		switch {
		case errors.Is(err, context.DeadlineExceeded):
			c.JSON(504, gin.H{"error": "timeout"})
		case errors.Is(err, ErrSubscriptionNotFound):
			c.JSON(404, gin.H{"error": "subscriptions not found"})
		default:
			h.logger.Error("failed to list subscriptions",
				zap.Int("status", 500),
				zap.String("error", err.Error()),
				zap.String("path", c.Request.URL.Path))
			c.JSON(500, gin.H{"error": "internal error"})
		}
		return
	}

	h.logger.Info("subscriptions listed successfully",
		zap.Int("count", response.Total),
		zap.Bool("has more", response.HasMore),
		zap.Int("offset", offset),
		zap.Int("limit", limit))

	c.JSON(200, response)
}

// GetTotalCost godoc
// @Summary      Calculate total cost of subscriptions
// @Description  Calculate total cost of all subscriptions in a given period with optional filters
// @Tags         subscriptions
// @Accept       json
// @Produce      json
// @Param        start_date   query  string  true  "Start date (MM-YYYY)"
// @Param        end_date     query  string  true  "End date (MM-YYYY)"
// @Param        user_id      query  string  false "Filter by user ID (UUID)"
// @Param        service_name query  string  false "Filter by service name"
// @Success      200  {object}  TotalCostResp
// @Failure      400  {object}  ErrorResponse  "Invalid parameters or date format"
// @Failure      500  {object}  ErrorResponse  "Internal server error"
// @Router       /subscriptions/total-cost [get]
func (h *Handler) GetTotalCost(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), requestTimeout)
	defer cancel()

	var req TotalCostReq
	if err := c.ShouldBindQuery(&req); err != nil {
		h.logger.Warn("invalid query parameters", zap.Error(err))
		c.JSON(400, gin.H{"error": "invalid query parameters"})
		return
	}

	response, err := h.service.CalculateTotalCost(ctx, req)
	if err != nil {
		switch {
		case errors.Is(err, context.DeadlineExceeded):
			c.JSON(504, gin.H{"error": "timeout"})
		case errors.Is(err, ErrSubscriptionNotFound):
			c.JSON(404, gin.H{"error": "subscriptions not found"})
		case errors.Is(err, ErrInvalidDateFormat), errors.Is(err, ErrEndDateBeforeStart):
			c.JSON(400, gin.H{"error": err.Error()})
		default:
			h.logger.Error("failed to calculate cost",
				zap.Int("status", 500),
				zap.String("error", err.Error()),
				zap.String("path", c.Request.URL.Path))
			c.JSON(500, gin.H{"error": "internal error"})
		}
		return
	}

	h.logger.Info("cost retrieved successfully",
		zap.Int("total_cost", response.TotalCost),
		zap.Int("subscriptions_count", response.SubscriptionsCount))

	c.JSON(200, response)
}

func getID(c *gin.Context) (int64, error) {
	idStr := c.Param("id")
	return strconv.ParseInt(idStr, 10, 64)
}